kids_content_engine/requirements.txt
g4f
edge-tts
moviepy
requests
pillow
python-dotenv
arabic-reshaper
python-bidi
numpy
aiohttp
curl_cffi
google-generativeai
